feat: store access tokens in access-tokens.conf
Instead of updating the nix.conf file constantly, store the tokens in a
separate file, and use the !include feature to load it from the main
config file.

If access-tokens exist in nix.conf, the tool will automatically migrate
them to the separate access-tokens.conf file.

That will reduce the number of backups we have to do. Plus it allows the
main config file to retain 0644 mode while keeping the access-tokens
locked down further.